[SPARK-46407][PS][TESTS] Reorganize `OpsOnDiffFramesDisabledTests`
### What changes were proposed in this pull request?
Reorganize `OpsOnDiffFramesDisabledTests`, factor it from the large `test_parity_ops_on_diff_frames.py`

### Why are the changes needed?
test code clean up

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44354 from zhengruifeng/ps_test_diff_frame_ops.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
